Final refactoring - Deliver currency
